Google Sound Search actions: Let's extend it
Because Google it's not the most comfortable way to use sound detection and maybe users want to use something more reliable or that saves their playlist on the cloud or other services. So let's call first to the most known apps and then rely on google app.

All the intents were taken from:
https://joaoapps.com/autoshare/intents/

(Thanks joaomgcd for the intents)